video: adaptive-quality fixes from the final review
- Gate goodput-driven steering on actual utilization (link-limited), not
  just the byte count: a scheduled bandwidth window measures bytes
  actually sent, not link capacity, so an idle desktop's tiny sample
  would otherwise walk quality all the way down and take ~21s to climb
  back. VideoStream now tracks the largest number of frames
  simultaneously outstanding since the last sample
  (maxPendingFramesSinceSample, updated from sendFrame() and
  onFrameAcknowledge()); AdaptiveQuality::step() only trusts the
  goodput-derived target when that backlog reached 2+, otherwise it aims
  for the cap (still a slow +5 climb) unless RTT says congested.
- Fix the RDPGFX_H264_QUANT_QUALITY metablock: its member order is
  {qpVal, qualityVal, qp, r, p} (freerdp/channels/rdpgfx.h), so the
  positional aggregate init `{qp, 0, currentQuality}` put the QP in the
  unused qpVal slot and left qualityVal at 0. Now sets .qp and
  .qualityVal by name.
- Fix a cap-raise/adaptive-quality desync: SessionController called
  session->setVideoQuality() directly alongside
  videoStream()->setQualityCap(), which (with adaptive quality on and
  the cap raised above the current, already-lower quality) forced the
  encoder to the new cap while VideoStream's own d->quality stayed at
  the old value - the two never resynced. SessionController now only
  calls setVideoQuality() directly when adaptive quality is off;
  VideoStream::setQualityCap() is the sole path otherwise, and now
  always emits requestedQualityChanged() (even when the computed value
  is unchanged) so a brand-new session still learns its quality.
- Tighten the RTT congestion gate: Input::averageRtt/minimumRtt are now
  microseconds instead of milliseconds, since truncating to milliseconds
  made 1-4 ms Wi-Fi jitter (e.g. min 1 ms, average 2 ms) look like a 2x
  spike. congested now additionally requires a real gap (>= 5 ms) above
  the minimum before the 1.5x ratio check.
- NetworkDetection::bandwidth() now computes in 64 bits before narrowing
  to kbit/s (the 32-bit multiply overflowed above ~4.3 Gbit/s).
- Correct several comments that misattributed which thread
  updateAdaptiveQuality()/setQualityCap()/requestedQualityChanged() run
  on: VideoStream is constructed on and never leaves the main thread, so
  NetworkDetection::bandwidthChanged (emitted from the FreeRDP peer
  thread) reaches it via a queued connection, not a direct one; the
  atomics on VideoStream's quality state are belt-and-braces, not load-
  bearing for a real cross-thread race today. No connection types were
  changed.

AdaptiveQualityTest: 11 passed (was 8; three new cases for the
link-limited gate and the RTT jitter fix).

+++ src/NetworkDetection.cpp
@@ -121,7 +121,9 @@ std::chrono::system_clock::duration NetworkDetection::averageRTT() const
 
 quint32 NetworkDetection::bandwidth() const
 {
-    return d->averageBandwidthBps.load() * 8 / 1000;
+    // Compute in 64 bits before narrowing back to kbit/s: averageBandwidthBps
+    // is bytes/s, and multiplying by 8 in 32 bits overflows above ~4.3 Gbit/s.
+    return static_cast<quint32>(static_cast<quint64>(d->averageBandwidthBps.load()) * 8 / 1000);
 }
 
 int NetworkDetection::validBandwidthSamples() const
